event form style similar to group form
